refactor(openomni): extract run-lifecycle from subagent runtime
