+++ packages/openomni/src/subagent/runtime.ts
@@ -2,11 +2,19 @@ import { ChatAgent, type ChatAgentConfig } from "@openomni/agent";
 import { type Guardrail, type Message, Subagent } from "@openomni/protocol";
 import { Bus, type BusEvent, Session, WorkerRun, type WorkerRunRecord } from "@openomni/session";
 import {
-  abort as abortSession,
   get as getAbortEntry,
   register as registerAbortController,
   remove as removeAbortController,
 } from "./abort-registry";
+import {
+  buildAbortSignal,
+  clearRunTimeouts,
+  finalizeRun,
+  raceAbortCompletion,
+  setupRunTimeouts,
+  shouldSkipFailureUpdate,
+  isTerminalStatus,
+} from "./run-lifecycle";
 import { sendToMailbox } from "./session-mailbox.js";
 import {
   type RuntimeModel,
@@ -132,156 +140,6 @@ async function runWithTranscript(
   });
 
   return agent.run({ messages });
-}
-
-function buildAbortSignal(controller: AbortController, signal?: AbortSignal): AbortSignal {
-  return AbortSignal.any(
-    [controller.signal, signal].filter(
-      (candidate): candidate is AbortSignal => candidate !== undefined,
-    ),
-  );
-}
-
-async function shouldSkipFailureUpdate(sessionId: string, runId: string): Promise<boolean> {
-  const run = await WorkerRun.get(sessionId, runId);
-  if (run?.status === "cancelled" || run?.status === "interrupted") return true;
-  // Cancel in progress: controller aborted but entry kept for completion tracking
-  const entry = getAbortEntry(sessionId, runId);
-  return !!entry && entry.controller.signal.aborted;
-}
-
-function isTerminalStatus(status: string): boolean {
-  return (
-    status === "succeeded" ||
-    status === "failed" ||
-    status === "cancelled" ||
-    status === "interrupted"
-  );
-}
-
-function resolveMetaStatus(runStatus: string | undefined): string {
-  switch (runStatus) {
-    case "succeeded":
-      return "idle";
-    case "failed":
-    case "cancelled":
-    case "interrupted":
-      return runStatus;
-    default:
-      return "idle";
-  }
-}
-
-async function finalizeRun(sessionId: string, runId: string): Promise<void> {
-  const run = await WorkerRun.get(sessionId, runId);
-  if (run && !isTerminalStatus(run.status) && !(await shouldSkipFailureUpdate(sessionId, runId))) {
-    await WorkerRun.updateStatus(sessionId, runId, "failed", { endedAt: Date.now() });
-    publishEvent(Subagent.Events.WorkerRunFailed, {
-      sessionId,
-      runId,
-      error: "non-terminal status at finally cleanup",
-    });
-  }
-
-  const finalRun = await WorkerRun.get(sessionId, runId);
-  const meta = Session.getWorkerMeta(sessionId);
-  if (meta && typeof meta === "object") {
-    Session.updateWorkerMeta(sessionId, {
-      ...meta,
-      status: resolveMetaStatus(finalRun?.status),
-    });
-  }
-}
-
-async function waitForAbortEntryRemoval(sessionId: string, runId: string): Promise<void> {
-  for (let i = 0; i < 20; i++) {
-    if (getAbortEntry(sessionId, runId) === undefined) return;
-    await Promise.resolve();
-  }
-  while (getAbortEntry(sessionId, runId) !== undefined) {
-    await new Promise((resolve) => setTimeout(resolve, 50));
-  }
-}
-
-async function raceAbortCompletion(
-  sessionId: string,
-  runId: string,
-  hardTimeoutMs: number,
-): Promise<void> {
-  let timeoutId: ReturnType<typeof setTimeout> | undefined;
-
-  const abortWait = waitForAbortEntryRemoval(sessionId, runId).then(() => "settled" as const);
-  const timeout = new Promise<"timeout">((resolve) => {
-    timeoutId = setTimeout(() => resolve("timeout"), hardTimeoutMs);
-  });
-
-  const winner = await Promise.race([abortWait, timeout]);
-
-  if (winner === "timeout") {
-    removeAbortController(sessionId, runId);
-    const run = await WorkerRun.get(sessionId, runId);
-    if (run && !isTerminalStatus(run.status)) {
-      await WorkerRun.updateStatus(sessionId, runId, "interrupted", { endedAt: Date.now() });
-    }
-    publishEvent(Subagent.Events.WorkerRunFailed, {
-      sessionId,
-      runId,
-      error: "cancel timeout exceeded",
-    });
-  } else {
-    clearTimeout(timeoutId);
-    const run = await WorkerRun.get(sessionId, runId);
-    if (run && !isTerminalStatus(run.status)) {
-      await WorkerRun.updateStatus(sessionId, runId, "cancelled", { endedAt: Date.now() });
-    }
-  }
-}
-
-type TimeoutTimers = {
-  soft?: ReturnType<typeof setTimeout>;
-  hard?: ReturnType<typeof setTimeout>;
-};
-
-function setupRunTimeouts(
-  sessionId: string,
-  runId: string,
-  softTimeoutMs?: number,
-  hardTimeoutMs?: number,
-): TimeoutTimers {
-  const timers: TimeoutTimers = {};
-
-  if (softTimeoutMs !== undefined) {
-    timers.soft = setTimeout(() => {
-      publishEvent(Subagent.Events.WorkerRunFailed, {
-        sessionId,
-        runId,
-        error: "soft timeout exceeded",
-      });
-    }, softTimeoutMs);
-  }
-
-  if (hardTimeoutMs !== undefined) {
-    timers.hard = setTimeout(async () => {
-      try {
-        await WorkerRun.updateStatus(sessionId, runId, "interrupted", { endedAt: Date.now() });
-      } catch {
-        return;
-      }
-      publishEvent(Subagent.Events.WorkerRunFailed, {
-        sessionId,
-        runId,
-        error: "hard timeout exceeded",
-      });
-      abortSession(sessionId, runId);
-    }, hardTimeoutMs);
-  }
-
-  return timers;
-}
-
-function clearRunTimeouts(timers: TimeoutTimers): void {
-  if (timers.soft) clearTimeout(timers.soft);
-  if (timers.hard) clearTimeout(timers.hard);
 }
 
 export namespace SubagentRuntime {
